Separate Route Files
Creates separate route file for each resource.
