+++ src/routers/task.js
@@ -1,0 +1,72 @@
+const express = require('express');
+const router = new express.Router();
+const Task = require('../models/task');
+
+router.post('/tasks', async(req, res) => {
+  const task = new Task(req.body);
+
+  try {
+    await task.save();
+    res.status(201).send(task);
+  } catch (e) {
+    res.status(422).send(e);
+  }
+});
+
+router.get('/tasks', async(req, res) => {
+  try {
+    const tasks = await Task.find({});
+    res.send(tasks);
+  } catch (e) {
+    res.status(500).send();
+  }
+});
+
+router.get('/tasks/:id', async(req, res) => {
+  try {
+    const task = await Task.findById(req.params.id);
+    if (!task) {
+      return res.status(404).send();
+    }
+    res.send(task);
+  } catch (e) {
+    res.status(500).send(e)
+  }
+});
+
+router.patch('/tasks/:id', async(req, res) => {
+  const updates = Object.keys(req.body);
+  const allowedUpdates = ['description', 'completed'];
+  const isValidOperation = updates.every((update) => allowedUpdates.includes(update));
+
+  if (!isValidOperation) {
+    res.status(400).send( { errors: 'Invalid updates!' });
+  }
+
+  try {
+    const task = await Task.findByIdAndUpdate(req.params.id, req.body, { new: true, runValidators: true });
+
+    if (!task) {
+      return res.status(404).send();
+    }
+
+    res.send(task);
+  } catch (e) {
+    res.status(400).send(e);
+  }
+});
+
+router.delete('/tasks/:id', async(req, res) => {
+  try {
+    const task = await Task.findByIdAndDelete(req.params.id);
+
+    if (!task) {
+      return res.status(404).send();
+    }
+    res.send(task);
+  } catch (e) {
+    res.status(500).send();
+  }
+});
+
+module.exports = router;
